refactor(dashboard): migrate small panels — overview / tools / permissions / plans / hooks
Stage 4 PR 4.1 from #28. Five panels move out of dashboard/app.js:

- panels/overview.ts (53 LoC) — Live Cockpit metric grid
- panels/tools.ts (71 LoC) — registered-tools table
- panels/permissions.ts (169 LoC) — allowlist CRUD + builtin
  groupings; private groupByVerb helper folded in
- panels/plans.ts (104 LoC) — archived-plan browser + step view
- panels/hooks.ts (105 LoC) — JSON editor for project + global hooks

Plus shared pieces these panels need:

- components/metric-card.ts (23 LoC) — used by overview now and
  by system in a later PR
- lib/html.ts (4 LoC) — single htm.bind(h) export so panels
  don't each rebind

dashboard/app.js loses 419 LoC. Bundle holds at ~122 KB; CDN
imports stay external. Typecheck + lint + 1682 tests still
pass. No behavior change.
