+++ dashboard/app.js
@@ -10,6 +10,7 @@ import {
   useRef,
   useState,
 } from "https://esm.sh/preact@10.22.0/hooks";
+import { MetricCard } from "./src/components/metric-card";
 import { MODE, TOKEN, api } from "./src/lib/api";
 import { ToastStack, appBus, showToast, toastBus } from "./src/lib/bus";
 import { ErrorBoundary, ErrorOverlay } from "./src/lib/error-boundary";
@@ -24,61 +25,15 @@ import {
   renderUnifiedDiff,
 } from "./src/lib/markdown";
 import { usePoll } from "./src/lib/use-poll";
+import { HooksPanel } from "./src/panels/hooks";
+import { OverviewPanel } from "./src/panels/overview";
+import { PermissionsPanel } from "./src/panels/permissions";
+import { PlansPanel } from "./src/panels/plans";
+import { ToolsPanel } from "./src/panels/tools";
 
 const html = htm.bind(h);
 
 // ---------- panels ----------
-
-function OverviewPanel() {
-  const { data, error, loading } = usePoll("/overview", 2000);
-  if (loading && !data) return html`<div class="boot">loading overview…</div>`;
-  if (error) return html`<div class="notice err">overview failed: ${error.message}</div>`;
-  const o = data;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Live Cockpit</h2>
-        <span class="panel-subtitle">${o.mode === "attached" ? "attached to running session" : "standalone (read-only disk view)"}</span>
-      </div>
-
-      ${o.mode === "standalone" ? html`<div class="notice">Standalone mode — start <code>/dashboard</code> from inside <code>reasonix code</code> for live session state, MCP, and tools.</div>` : null}
-
-      <div class="metric-grid">
-        ${MetricCard("Reasonix", o.version, o.latestVersion && o.latestVersion !== o.version ? `latest: ${o.latestVersion}` : "current")}
-        ${MetricCard("Session", o.session ?? "—", o.session === null ? "ephemeral or disconnected" : null)}
-        ${MetricCard("Model", o.model ?? "—", o.model ? "active" : null)}
-        ${MetricCard("Edit mode", o.editMode ?? "—", o.editMode === "yolo" ? "all prompts bypassed" : null, o.editMode === "yolo" ? "warn" : null)}
-        ${MetricCard("Plan mode", o.planMode === null ? "—" : o.planMode ? "ON" : "off", o.planMode ? "writes gated" : null)}
-        ${MetricCard("Pending edits", fmtNum(o.pendingEdits), o.pendingEdits ? "awaiting /apply" : null)}
-        ${MetricCard("MCP servers", fmtNum(o.mcpServerCount), null)}
-        ${MetricCard("Tools", fmtNum(o.toolCount), null)}
-      </div>
-
-      <div class="section-title">Working directory</div>
-      <div class="card">
-        <div class="card-value mono" style="font-size: 14px;">${o.cwd ?? "—"}</div>
-      </div>
-    </div>
-  `;
-}
-
-function MetricCard(title, value, hint, pillVariant) {
-  const muted = value === "—" || value === null || value === undefined;
-  return html`
-    <div class="card">
-      <div class="card-title">${title}</div>
-      <div class="card-value ${muted ? "muted" : ""}">${value}</div>
-      ${
-        hint
-          ? pillVariant
-            ? html`<div class="card-hint"><span class="pill pill-${pillVariant}">${hint}</span></div>`
-            : html`<div class="card-hint">${hint}</div>`
-          : null
-      }
-    </div>
-  `;
-}
 
 function UsagePanel() {
   const { data, error, loading } = usePoll("/usage", 5000);
@@ -166,209 +121,6 @@ function UsagePanel() {
     </div>
   `;
 }
-
-function ToolsPanel() {
-  const { data, error, loading } = usePoll("/tools", 4000);
-  if (loading && !data) return html`<div class="boot">loading tools…</div>`;
-  if (error?.status === 503) {
-    return html`<div class="notice">${error.body?.error ?? "live tools view requires an attached session"}</div>`;
-  }
-  if (error) return html`<div class="notice err">tools failed: ${error.message}</div>`;
-  const t = data;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Tools</h2>
-        <span class="panel-subtitle">${t.total} registered ${t.planMode ? html`<span class="pill pill-warn">plan mode — writes gated</span>` : ""}</span>
-      </div>
-      ${
-        t.tools.length === 0
-          ? html`<div class="empty">No tools registered.</div>`
-          : html`
-          <table>
-            <thead>
-              <tr>
-                <th>name</th>
-                <th>flags</th>
-                <th>description</th>
-              </tr>
-            </thead>
-            <tbody>
-              ${t.tools.map(
-                (tool) => html`
-                <tr>
-                  <td><code>${tool.name}</code></td>
-                  <td>
-                    ${tool.readOnly ? html`<span class="pill pill-ok">read-only</span>` : html`<span class="pill pill-accent">write</span>`}
-                    ${tool.flattened ? html` <span class="pill pill-dim">flat</span>` : ""}
-                  </td>
-                  <td>${tool.description ?? ""}</td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-    </div>
-  `;
-}
-
-function PermissionsPanel() {
-  const { data, error, loading, refresh } = usePoll("/permissions", 5000);
-  const [draft, setDraft] = useState("");
-  const [busy, setBusy] = useState(false);
-  const [feedback, setFeedback] = useState(null);
-
-  const add = useCallback(async () => {
-    const prefix = draft.trim();
-    if (!prefix) return;
-    setBusy(true);
-    setFeedback(null);
-    try {
-      const res = await api("/permissions", { method: "POST", body: { prefix } });
-      if (res.alreadyPresent) setFeedback({ kind: "info", text: `${prefix} already in list` });
-      else setFeedback({ kind: "ok", text: `added: ${prefix}` });
-      setDraft("");
-      await refresh();
-    } catch (err) {
-      setFeedback({ kind: "err", text: err.message });
-    } finally {
-      setBusy(false);
-    }
-  }, [draft, refresh]);
-
-  const remove = useCallback(
-    async (prefix) => {
-      if (!confirm(`Remove "${prefix}" from this project's allowlist?`)) return;
-      setBusy(true);
-      setFeedback(null);
-      try {
-        await api("/permissions", { method: "DELETE", body: { prefix } });
-        setFeedback({ kind: "ok", text: `removed: ${prefix}` });
-        await refresh();
-      } catch (err) {
-        setFeedback({ kind: "err", text: err.message });
-      } finally {
-        setBusy(false);
-      }
-    },
-    [refresh],
-  );
-
-  const clearAll = useCallback(async () => {
-    if (!confirm("Wipe every project allowlist entry? Builtin entries are unaffected.")) return;
-    setBusy(true);
-    setFeedback(null);
-    try {
-      const res = await api("/permissions/clear", { method: "POST", body: { confirm: true } });
-      setFeedback({
-        kind: "ok",
-        text: `cleared ${res.dropped} entr${res.dropped === 1 ? "y" : "ies"}`,
-      });
-      await refresh();
-    } catch (err) {
-      setFeedback({ kind: "err", text: err.message });
-    } finally {
-      setBusy(false);
-    }
-  }, [refresh]);
-
-  if (loading && !data) return html`<div class="boot">loading permissions…</div>`;
-  if (error) return html`<div class="notice err">permissions failed: ${error.message}</div>`;
-  const p = data;
-
-  const banner =
-    p.editMode === "yolo"
-      ? html`<div class="notice warn">YOLO mode — every shell command auto-runs, allowlist bypassed. <code>/mode review</code> in TUI re-enables.</div>`
-      : null;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Permissions</h2>
-        <span class="panel-subtitle">${p.currentCwd ? `project: ${p.currentCwd}` : "no active project"}</span>
-      </div>
-      ${banner}
-
-      ${
-        p.currentCwd
-          ? html`
-          <div class="row">
-            <input
-              type="text"
-              placeholder='add a prefix, e.g. "npm run build" or "deploy.sh"'
-              value=${draft}
-              onInput=${(e) => setDraft(e.target.value)}
-              onKeyDown=${(e) => {
-                if (e.key === "Enter") add();
-              }}
-              disabled=${busy}
-            />
-            <button class="primary" onClick=${add} disabled=${busy || !draft.trim()}>Add</button>
-            <button class="danger" onClick=${clearAll} disabled=${busy || p.project.length === 0}>Clear all</button>
-          </div>
-          ${feedback ? html`<div class="notice ${feedback.kind === "err" ? "err" : feedback.kind === "ok" ? "" : "warn"}">${feedback.text}</div>` : null}
-        `
-          : html`<div class="notice">Mutations require <code>/dashboard</code> from inside an active <code>reasonix code</code> session — standalone <code>reasonix dashboard</code> can't tell which project's allowlist to edit.</div>`
-      }
-
-      <div class="section-title">Project allowlist (${p.project.length})</div>
-      ${
-        p.project.length === 0
-          ? html`<div class="empty">Nothing stored yet for this project.</div>`
-          : html`
-          <table>
-            <thead><tr><th>#</th><th>prefix</th><th></th></tr></thead>
-            <tbody>
-              ${p.project.map(
-                (prefix, i) => html`
-                <tr>
-                  <td class="muted">${i + 1}</td>
-                  <td><code>${prefix}</code></td>
-                  <td class="numeric">${p.currentCwd ? html`<button class="danger" onClick=${() => remove(prefix)} disabled=${busy}>remove</button>` : null}</td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-
-      <div class="section-title">Builtin allowlist (${p.builtin.length}) — read-only, baked in</div>
-      <div class="card mono" style="font-size: 12px; line-height: 1.7;">
-        ${groupByVerb(p.builtin).map(
-          ([verb, list]) => html`
-          <div><span class="pill pill-dim">${verb}</span> ${list.join(" · ")}</div>
-        `,
-        )}
-      </div>
-    </div>
-  `;
-}
-
-function groupByVerb(list) {
-  const groups = new Map();
-  for (const entry of list) {
-    const head = entry.split(" ")[0];
-    if (!groups.has(head)) groups.set(head, []);
-    const tail = entry.slice(head.length).trim();
-    groups.get(head).push(tail || "(bare)");
-  }
-  return [...groups.entries()];
-}
-
-// ---------- Chat panel ----------
-
-const ROLE_GLYPH = {
-  user: "◇",
-  assistant: "◆",
-  tool: "▣",
-  info: "·",
-  warning: "▲",
-  error: "✦",
-};
 
 function renderMessageBody(text) {
   if (!text) return null;
@@ -1812,84 +1564,6 @@ function SessionsPanel() {
 
 // ---------- Plans archive ----------
 
-function PlansPanel() {
-  const { data, error, loading } = usePoll("/plans", 8000);
-  const [open, setOpen] = useState(null);
-  if (loading && !data) return html`<div class="boot">loading plans…</div>`;
-  if (error) return html`<div class="notice err">plans failed: ${error.message}</div>`;
-  const plans = data.plans ?? [];
-
-  if (open) {
-    const completedSet = new Set(open.completedStepIds);
-    return html`
-      <div>
-        <div class="panel-header">
-          <h2 class="panel-title">Plan</h2>
-          <span class="panel-subtitle">${open.session} · ${fmtRelativeTime(open.completedAt)}</span>
-          <button onClick=${() => setOpen(null)} style="margin-left: auto;">← back</button>
-        </div>
-        ${open.summary ? html`<div class="notice">${open.summary}</div>` : null}
-        <div class="card">
-          ${open.steps.map((step) => {
-            const done = completedSet.has(step.id);
-            return html`
-              <div style="padding: 8px 0; border-bottom: 1px solid var(--border); display: flex; gap: 12px;">
-                <div style="width: 16px; color: ${done ? "var(--ok)" : "var(--fg-3)"}; font-family: var(--mono);">
-                  ${done ? "✓" : "·"}
-                </div>
-                <div style="flex: 1;">
-                  <div style="color: ${done ? "var(--fg-2)" : "var(--fg-0)"}; font-weight: 500;">
-                    ${step.title}
-                  </div>
-                  ${step.action ? html`<div style="color: var(--fg-2); font-size: 12px; margin-top: 2px;">${step.action}</div>` : null}
-                  ${step.risk ? html`<span class="pill pill-${step.risk === "high" ? "err" : step.risk === "medium" ? "warn" : "dim"}" style="margin-top: 4px;">${step.risk}</span>` : null}
-                </div>
-              </div>
-            `;
-          })}
-        </div>
-      </div>
-    `;
-  }
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Plans</h2>
-        <span class="panel-subtitle">${plans.length} archived · click to view</span>
-      </div>
-      ${
-        plans.length === 0
-          ? html`<div class="empty">No archived plans yet — run a turn that calls <code>submit_plan</code> + <code>mark_step_complete</code>.</div>`
-          : html`
-          <table>
-            <thead>
-              <tr>
-                <th>session</th>
-                <th>title</th>
-                <th class="numeric">progress</th>
-                <th class="numeric">archived</th>
-              </tr>
-            </thead>
-            <tbody>
-              ${plans.map(
-                (p, i) => html`
-                <tr key=${i} onClick=${() => setOpen(p)} style="cursor: pointer;">
-                  <td><code>${p.session}</code></td>
-                  <td>${p.summary ?? html`<span class="muted">(no title)</span>`}</td>
-                  <td class="numeric">${p.completedSteps}/${p.totalSteps} · ${fmtPct(p.completionRatio)}</td>
-                  <td class="numeric muted">${fmtRelativeTime(p.completedAt)}</td>
-                </tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-      }
-    </div>
-  `;
-}
-
 // ---------- Usage time-series chart (uPlot) ----------
 
 let uPlotPromise = null;
@@ -2260,93 +1934,6 @@ function SettingsPanel() {
 }
 
 // ---------- Hooks ----------
-
-function HooksPanel() {
-  const [data, setData] = useState(null);
-  const [error, setError] = useState(null);
-  const [drafts, setDrafts] = useState({}); // { project: jsonText, global: jsonText }
-  const [busy, setBusy] = useState(false);
-  const [info, setInfo] = useState(null);
-
-  const load = useCallback(async () => {
-    try {
-      const r = await api("/hooks");
-      setData(r);
-      setDrafts({
-        project: JSON.stringify(r.project.hooks ?? {}, null, 2),
-        global: JSON.stringify(r.global.hooks ?? {}, null, 2),
-      });
-    } catch (err) {
-      setError(err.message);
-    }
-  }, []);
-  useEffect(() => {
-    load();
-  }, [load]);
-
-  const saveScope = useCallback(
-    async (scope) => {
-      setBusy(true);
-      setError(null);
-      let parsed;
-      try {
-        parsed = JSON.parse(drafts[scope]);
-      } catch (err) {
-        setError(`${scope} JSON: ${err.message}`);
-        setBusy(false);
-        return;
-      }
-      try {
-        await api("/hooks/save", { method: "POST", body: { scope, hooks: parsed } });
-        await api("/hooks/reload", { method: "POST", body: {} });
-        setInfo(`saved + reloaded ${scope}`);
-        setTimeout(() => setInfo(null), 3000);
-        await load();
-      } catch (err) {
-        setError(err.message);
-      } finally {
-        setBusy(false);
-      }
-    },
-    [drafts, load],
-  );
-
-  if (!data && !error) return html`<div class="boot">loading hooks…</div>`;
-  if (error && !data) return html`<div class="notice err">${error}</div>`;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Hooks</h2>
-        <span class="panel-subtitle">${data.resolved.length} resolved · events: ${data.events.join(", ")}</span>
-      </div>
-      ${info ? html`<div class="notice">${info}</div>` : null}
-      ${error ? html`<div class="notice err">${error}</div>` : null}
-      ${["project", "global"].map((scope) => {
-        const meta = data[scope];
-        return html`
-          <div class="section-title">${scope} — <code>${meta.path ?? "(no project)"}</code></div>
-          ${
-            scope === "project" && !meta.path
-              ? html`<div class="empty">No active project — open <code>/dashboard</code> from <code>reasonix code</code> to edit project hooks.</div>`
-              : html`
-              <textarea
-                style="width: 100%; height: 240px; font-family: var(--mono); font-size: 12.5px; background: var(--bg-2); color: var(--fg-0); border: 1px solid var(--border); border-radius: var(--radius-md); padding: 10px;"
-                value=${drafts[scope] ?? ""}
-                onInput=${(e) => setDrafts({ ...drafts, [scope]: e.target.value })}
-                disabled=${busy}
-              ></textarea>
-              <div class="row" style="margin-top: 8px;">
-                <button class="primary" disabled=${busy} onClick=${() => saveScope(scope)}>Save + Reload</button>
-                <button disabled=${busy} onClick=${load}>Discard changes</button>
-              </div>
-            `
-          }
-        `;
-      })}
-    </div>
-  `;
-}
 
 // ---------- Memory ----------
