+++ dashboard/src/lib/html.ts
@@ -1,0 +1,4 @@
+import htm from "https://esm.sh/htm@3.1.1";
+import { h } from "https://esm.sh/preact@10.22.0";
+
+export const html = htm.bind(h);
